client_id wasn't found any more. Soundcloud has changed the syntax for the client_id in the .js-File so the regular expression which searches for the client_id had to be changed.
Also the client_id seems to be stored in different .js files, now. In
the past the filename with the id had a certain pattern which could be
searched for.
Changed the get_credentials() function and its
"helper"-function find_script_url() so that the js-files iteratably
are getting searched for the client-id until it is found.

flomet